break interface compatibility with an option to preserve existing behavior
Previously, we always supplied the request object as a private _request key
and expected the response to be the full object returned by cue. In addition
the response is expected to be a State message for the desired state which
does not allow the composition to set the context, for example.

This changes the interface such that, by default:

* the request object is sent as a `request` key without the leading underscore
* the response object is expected to be a `response` key in the output
* the response message is now a RunFunctionResponse rather than a State message

Backward-compatible knobs are provided to get the previous behavior.

This allows composition authors to get rid of the underscore in front of all
keys that should not be returned and have cleaner code. It allows allows
a cue composition to set values in the context.